requirements.txt
fastapi==0.111.0
pydantic==2.7.1
pydantic_core==2.18.2
motor==3.3.1
setuptools~=68.2.0
python-dotenv~=1.0.1
SQLAlchemy~=2.0.30